Updates to Competition Sketch and Library
Filled in Competition Sketch.
Created new system to use different power levels in the solenoid
depending on distance. This requires new landingDistanceIdeal functions,
one per power level: HIGH(255) and LOW(240).
High is used for distances farther than 109 cm, low is used for anything
closer. The purpose of this is to have a deep landing angle for all
targets within range. Coefficients for the polynomial models will come
in the following commit.