Updates Competition Sketch & moveTo Function
Added a number of comments to document the purpose of many of the
commands, and reorganized the ordering of some variable declarations to
minimize unnecessary memory allocation.

I added an extra parameter to the moveTo function. It is a boolean named
stop. If true, after moving to the requested location, the cannon will
break to land in the middle of the color bar. If false, the cannon will
be left coasting. This is so that we can minimize unnecessary stopping
and waiting when moving the cannon servo to the calculated angle.

Also updated the model coefficients in the library to better represent
reality. These coefficients were calculated using the improved
optimizing scripts in MATLAB.